refactor: harden input.link deprecation parity
Mirror the output-side hardening on the input slot: non-optional readonly
declare, dedicated NodeInputSlot deprecation tests (replacing the combined
NodeSlot.deprecation file), and align the output.links assignment test with
the warn-and-ignore setter semantics. Update migration-plan status blocks
for the shipped input.link deletion.
